// src/order/types/enums/notification.enum.ts
export enum PharmacyActionType {
  BUCKET_CANCELLED_IN_PAYMENT = 'BUCKET_CANCELLED_IN_PAYMENT',
  BUCKET_CANCELLED_BY_PHARMACY_WITH_REASON = 'BUCKET_CANCELLED_BY_PHARMACY_WITH_REASON',
  USER_CANCEL_THE_BUCKET = 'USER_CANCEL_THE_BUCKET'
}

export enum PatientActionType {
  ORDER_CANCELLED = 'ORDER_CANCELLED'
}